functions no longer need a return type immediately passed into them (helps for building a functions later on.

+++ src/core.rs
@@ -1,4 +1,4 @@
 /// core VM data structures
 
-pub type Register = usize;
+pub type Register = i64;
 pub type RegList = Vec<Register>;
